Hamed and I worked towards user story #11. We created a API route that accepts a GET request to which it responds with the authorization status of Barclays from the FCA API.
The API key used to access it is retrieved from a local .env file.

I then led on implementing the test for our Given-When-Then scenario. Using Jest and Supertest. I decided to use supertest as it provides a suite of convieniance methods that make testing http routes easier.

I moved the starting of the express server to a different file. This lets supertest import the express app object without triggering a express instance to be launched that supertest can't close at the end of testing. Which would otherwise cause the command to hang indefinitely.

Co-authored-by: Jack Richards <richardsj31@cardiff.ac.uk>
Co-authored-by: Hamed Al Sarmi <alsarmih@cardiff.ac.uk>