Added a second api route to handle the individual logic, i was able to get a response back in boolean to check for "certified" and "Approved by regulator" as both of these are to return true.
During the next commit I will be habdling the logic to say whether an indivdual is "unathorized" by the firm then should check for individual. However the problem at the moment is to find a company example with such conditions in order to test. Or I will have to create a mock