Separated FCA-related business logic from the router file. Making the new aggregator component act as a 'middle-man' between them.
I have added a type used for the response returned by the aggregator to the method caller (the router in this case). Creating a strict expectation of the exact data to be returned.

Error-handling is included in the form of try-catch, some code comments regarding that are in this commit.

Unix timestamp is also provided as part of the response to provide some metadata which may be useful to the requestor.