Progress towards #23. I have created a new class called GamblingCommission an instance of which can be produced by its factory. It contains the logic needed to process either byte-representations of the GamblingCommission CSVs or a local CSV.
Making use of the COPY feature from Postgres to stream this new CSV data into two new tables I have created.

Code is commented throughout, jdocs are provided for most methods. A number of edge-cases are handled and the data provided goes through an amount of pre-processing to make it as valid as possible.

COPY process happens within a transaction, so if anything goes wrong the changes will not be committed to the database.