Completes #23. Users will be able to upload Gambling Commission CSVs to the database in a highly-performant manner.
Streams are used to process the provided CSV data in pieces instead of loading the entire CSV outright into memory.

An amount of pre-processing is performed on the CSVs. This is an area that can be expanded upon if needed in the future, if the need arises.

I make use of a factory to create a instance of the GamblingCommission class. In the process ensuring the instance has all it needs to function properly.

A level of validation is done on the provided file, making sure the field the file falls under matches one in our records. Also checking the mimetype of the file, rejecting it if it's not text/csv.

I have made two easy to use methods. One that processes a local CSV and uploads it to the relevant table. Which table and file to use is calculated through the key provided to the function.

This method will likely be called when another component has retrieved a new copy of the CSV and has it stored in the expected location.

The other function as touched upon before, takes a stream and pipes the CSV data into the database.

Both these methods run the integral database parts within a transaction. Meaning that if any errors occur during the process, all changes are not committed. Which is good because it means the existing data is not compromised upon an error.

I need to add more tests, but for now the one specified in the user story is complete.