For many days and hours I have been tackling one problem. The test I had created, and can be seen heavily edited in this commit was failing.
It wasn't failing due to any obvious reason, so I had to spend a very long time researching.

The database was, according to the test, not being not being populated with the new CSV data. Meanwhile, upon checking the database shortly afterwards, the data was all there.

So I thought it was a problem with jest not properly awaiting the asynchronous uploadCsv() method. To cut a long story short it wasn't.

Branching off this idea, I thought it might be because database transactions are not visible to other clients until the changes are committed. This also wasn't the issue.

I considered that it could be busboy causing the issue. Busboy doesn't immediately return anything upon the data being piped to it. Instead you have to rely on events emitted by it to continue your logic therein.

I thought that because of this jest may have been resolving the promise immediately upon the method being called. As there was nothing immediate to do. So, to remedy this, I tried wrapping the busboy logic in a promise, and awaiting that promise. But to no avail.

In the end, and I don't know why, but it seems like moving the mock Implementation I made for the uploadCsv method fixed the issue. The purpose of said mock was to intercept the call to uploadCsv and recall it with a different schema parameter (making it target the test schema).

I'm not sure why it worked, but at this point I'm happy with the result. I also experimented with another whole method of testing this component, which for the sake of my reflective report I have left commented. I'll remove any commented unused code later when it comes time to merge.