Completes #23. This story took a signficant amount of time, the mixing of asynchronous and event-driven code made this very challenging to test properly.
Working on this outlined a lot of hidden unexpected behaviours in Node.js. Like for instance, if you throw an error within a busboy event handler. It will not reject the asynchronous function that wraps it as you may expect. But instead it will hang indefnitely.

Regardless of the difficulties posed by this user story it is now complete. I have added tests for the rest of the acceptance criteria, all passing.

I have expanded upon the validation logic, making it even more resilient. To the point where you could forgo certain initial validation checks.

And the processing logic would still be able to reject the invalid data later on, retaining the original database state (via a transaction).

A full explanation of what has been done will be provided in the merge request.