Added the logic so if isAuthorised comes back as false then it would excetue for checking fo the individual. As of now we dont have a IRN nor dose the user pass it. This feature is supposed to help companies house in the future.
Ive also setted up a enviroment vairbale which is RUN_FCA_CHECK this would check the env file to see if the feature is required or not. Ive though of just leaving the logic commented out, but Carl mentioed this and I thought its a better way and more professianl to have it setted up as this. So if companies house or our developers dont want this logic to be excetured then they would hae it as false.

Progress towards user story #35