Completes #42 functionally. I renamed the Business Name Standardizer class to Business Name Processor. As I intend for it to be a more general class to handle tasks relating to business names.
In that class, I implemented a method that takes a business name (typically the one provided by the user), and the reference ID associated with a row under a regulatory body.

It then fetches the row in our records using this ID, and compares the given business name against the one associated with the record.

The provided name is standardized before this comparison is made. If they do not match this is reflected in the response the method gives. It will also provide a error message showcases the difference in this instance.

Otherwise, it will return 'true' indicating a match.

I plan to use this method within the aggregator to provide more context to the client if the business name doesn't exactly match in our systems.