Functionally completes #45. I now need to create the tests.
I have made a snapshot manager class that handles the logic behind creating and maintaining snapshots for database tables.

The expected way to use the class is to instantiate it with a connected database client that is operating within a transaction.

I have integrated it into the upload CSV flow, so that once all the data has been successfully uploaded to the appropriate table. It will, before the transactions changes are committed, create a snapshot of the table.

The 'snapshot' being a new table containing the data, timestamped in its name. If there are already 5 existing snapshots retained for the body, the manager will automatically delete the oldest one to make room for the newest snapshot.