Fixes a variety of issues found in the codebase while manually running through the flow.
1. Changed way the business name was being accessed from the DataRow for HMRC CSV data.

Uses row['Business Name'] instead of row.BUSINESS_NAME; matching the format of it within the actual object.

Fixes issues with HMRC upload process failing.

2. Snapshot manager was not correctly identifying old snapshot tables to delete past the limit of 5. This was because the name passed into the query being wrong.

I have now returned the actual basename of the table, and have constructed a proper snapshot table name from this.

Highlights how the testing for the Snapshot module may not accurately test for the deletion of old tables, past the default limit of 5.

Though, due to time constraints I don't think it's possible for me to rework the tests. A point for my reflective report I feel.

In short, oldest snapshot past the limit are now deleted.

3. One of the Scraping components methods was making an additional call to update using the Gambling Commission CSVs.

This was unintended, as it should only happen once on the initial load, not twice.

Now fixed so it will perform the update once.

4. I incorporated an anonymous async arrow function that will auto-execute on application load.

Making sure to use await, so the creation of the schema is properly awaited before moving onto the next set-up steps.